combine two back commands
